requirements.txt
atlassian-python-api==4.0.3
azure-functions==1.23.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
Deprecated==1.2.18
idna==3.10
jira==3.8.0
jmespath==1.0.1
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
pillow==11.2.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
slack_sdk==3.35.0
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
